% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{makepath}
\alias{makepath}
\title{Returns the path to the API URL for the requested service.}
\usage{
makepath(service)
}
\arguments{
\item{service}{The name of the service}
}
\value{
The path to the API service.
}
\description{
Returns the path to the API URL for the requested service.
}
\examples{
makepath("stations-v2")
}

