% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{my_sd}
\alias{my_sd}
\title{My standard deviation function}
\usage{
my_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Remove missing values?}
}
\description{
My standard deviation function
}
