% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_departures.R
\name{ns_departures}
\alias{ns_departures}
\title{Title}
\usage{
ns_departures(id)
}
\arguments{
\item{id}{The code (abbreviation) or short name or medium-length name or full name or synonym of the station's name}
}
\description{
Title
}

