% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_disruptions.R
\name{ns_disruptions}
\alias{ns_disruptions}
\title{Disruptions}
\usage{
ns_disruptions(id, actual = TRUE, unplanned = TRUE)
}
\arguments{
\item{id}{The code (abbreviation) or short name or medium-length name or full name or synonym of the station's name}

\item{unplanned}{Boolean (true or false) indicator of the scheduled engineering work for the next two weeks must be returned. This parameter provides the information described above in section 2. Note: unplanned=true will return the scheduled engineering work. This is therefore the opposite of what the parameter name would imply.}

\item{actualBoolean}{(true or false) indicator of the current disruptions must be returned. This includes both unscheduled disruptions at the moment of the request, as well as engineering work scheduled to take place within two hours of the request. This parameter will provide the information described above in section 1.}
}
\description{
Disruptions
}

