% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\docType{data}
\name{BASE_URL}
\alias{BASE_URL}
\alias{ns}
\title{Connect to NS API and return information about planned and current time tables.}
\format{An object of class \code{character} of length 1.}
\usage{
BASE_URL

ns(path, query = NULL)
}
\arguments{
\item{path}{}
}
\description{
NS requires users to register to access the API. Upon registration,
https://www.ns.nl/ews-aanvraagformulier/?0
   This is a description

Title
}
\seealso{
\link{http://www.ns.nl/en/travel-information/ns-api}

The URL to the webservices API
}
\keyword{datasets}

